move nested data utils to utils submodule
